Code cleanup

+++ openchrom/plugins/net.openchrom.msd.converter.supplier.mzmlb/src/net/openchrom/msd/converter/supplier/mzmlb/internal/io/ReaderProxy.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 Lablicate GmbH.
+ * Copyright (c) 2021, 2022 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -77,7 +77,7 @@ public class ReaderProxy implements IReaderProxy {
 
 		double[] output = new double[input.length];
 		for(int i = 0; i < input.length; i++) {
-			output[i] = (double)input[i];
+			output[i] = input[i];
 		}
 		return output;
 	}
